# Implement Two-Tiered Sampling for Optimal Performance and Accuracy