--- a/main.py
+++ b/main.py
@@ -101,12 +101,21 @@ def calculate_abtest(data: ABTestInput):
     # Заметка: переменные a_samples и b_samples будут созданы на следующем шаге плана.
     # На этом шаге мы только вычисляем alpha_a_posterior, beta_a_posterior, alpha_b_posterior, beta_b_posterior.
 
-    sampling_start_time = time.time()
-    N_SAMPLES = 20000 # Количество сэмплов для генерации
-    a_samples = beta.rvs(alpha_a_posterior, beta_a_posterior, size=N_SAMPLES, random_state=42)
-    b_samples = beta.rvs(alpha_b_posterior, beta_b_posterior, size=N_SAMPLES, random_state=42)
-    sampling_end_time = time.time()
-    print(f"[LOG] Sampling from posterior Beta distributions (N={N_SAMPLES}): {sampling_end_time - sampling_start_time:.4f}s")
+    N_SAMPLES_METRICS = 1000000
+    N_SAMPLES_PLOTS = 50000
+
+    sampling_metrics_start_time = time.time()
+    a_samples_full = beta.rvs(alpha_a_posterior, beta_a_posterior, size=N_SAMPLES_METRICS, random_state=42)
+    b_samples_full = beta.rvs(alpha_b_posterior, beta_b_posterior, size=N_SAMPLES_METRICS, random_state=42)
+    sampling_metrics_end_time = time.time()
+    print(f"[LOG] Sampling for metrics (N={N_SAMPLES_METRICS}): {sampling_metrics_end_time - sampling_metrics_start_time:.4f}s")
+
+    # Создание подвыборок для графиков
+    slicing_plot_samples_start_time = time.time()
+    a_samples_plot = a_samples_full[:N_SAMPLES_PLOTS]
+    b_samples_plot = b_samples_full[:N_SAMPLES_PLOTS]
+    slicing_plot_samples_end_time = time.time()
+    print(f"[LOG] Slicing samples for plots (N={N_SAMPLES_PLOTS}): {slicing_plot_samples_end_time - slicing_plot_samples_start_time:.4f}s")
 
     bayesian_end_time = time.time() # Завершение всего байесовского блока
     print(f"[LOG] Bayesian part (Analytic Beta): Total {bayesian_end_time - bayesian_start_time:.4f}s")
@@ -122,11 +131,8 @@ def calculate_abtest(data: ABTestInput):
     processing_start_time = time.time()
     print(f"[LOG] Data processing: Start")
 
-    # extract_samples_start_time = time.time() # Removed
-    # a_samples = trace.posterior['p_a'].values.flatten() # Removed
-    # b_samples = trace.posterior['p_b'].values.flatten() # Removed
-    # extract_samples_end_time = time.time() # Removed
-    # print(f"[LOG] Sample extraction: {extract_samples_end_time - extract_samples_start_time:.4f}s") # Removed
+    # extract_samples_start_time = time.time() # This block is removed as samples are now a_samples_full/plot, b_samples_full/plot
+    # print(f"[LOG] Sample extraction: ...") # This log is removed
 
     # The following lines will need new a_samples, b_samples or be removed/modified
     # For now, keeping them commented out or aware they will cause errors
@@ -137,69 +143,88 @@ def calculate_abtest(data: ABTestInput):
     # The following lines will need new a_samples, b_samples or be removed/modified
     # For now, providing dummy values or commenting out to prevent immediate errors
     # but expecting these to be handled by subsequent sample generation logic.
-    a_dist = []
-    b_dist = []
-    kde_calc_start_time = time.time()
-    a_kde = gaussian_kde(a_samples)
-    b_kde = gaussian_kde(b_samples)
-    a_dist = a_kde(x)
-    b_dist = b_kde(x)
-    if np.max(a_dist) > 0: a_dist = a_dist / np.max(a_dist)
-    if np.max(b_dist) > 0: b_dist = b_dist / np.max(b_dist)
-    kde_calc_end_time = time.time()
-    print(f"[LOG] KDE calculations (a_dist, b_dist): {kde_calc_end_time - kde_calc_start_time:.4f}s")
+    # a_dist = [] # These will be calculated below using _plot samples
+    # b_dist = [] # These will be calculated below using _plot samples
 
     bayes_metrics_start_time = time.time()
-    prob_b_better = float(np.mean(b_samples > a_samples))
-    prob_a_better = float(np.mean(a_samples > b_samples))
-    a_mean = float(np.mean(a_samples))
-    b_mean = float(np.mean(b_samples))
-    # added .any() for safety, though with Beta priors giving samples in (0,1) and non-zero success/total, samples should not be all zero.
-    a_expected_loss = float(np.mean(np.where(b_samples > a_samples, (b_samples - a_samples) / b_samples if b_samples.any() else 0, 0)))
-    b_expected_loss = float(np.mean(np.where(a_samples > b_samples, (a_samples - b_samples) / a_samples if a_samples.any() else 0, 0)))
+    # Используем _full сэмплы для метрик
+    prob_b_better = float(np.mean(b_samples_full > a_samples_full))
+    prob_a_better = float(np.mean(a_samples_full > b_samples_full))
+    a_mean = float(np.mean(a_samples_full))
+    b_mean = float(np.mean(b_samples_full))
+    # Для expected_loss также используем _full сэмплы
+    a_expected_loss = float(np.mean(np.where(b_samples_full > a_samples_full, (b_samples_full - a_samples_full) / b_samples_full if b_samples_full.any() else 0, 0)))
+    b_expected_loss = float(np.mean(np.where(a_samples_full > b_samples_full, (a_samples_full - b_samples_full) / a_samples_full if a_samples_full.any() else 0, 0)))
     bayes_metrics_end_time = time.time()
-    print(f"[LOG] Bayesian metrics (prob_b_better, means, expected_loss): {bayes_metrics_end_time - bayes_metrics_start_time:.4f}s")
+    print(f"[LOG] Bayesian metrics (from N={N_SAMPLES_METRICS} samples): {bayes_metrics_end_time - bayes_metrics_start_time:.4f}s")
+
+    kde_calc_start_time = time.time()
+    # Используем _plot сэмплы для KDE
+    if a_samples_plot.size > 1: a_kde = gaussian_kde(a_samples_plot)
+    else: a_kde = None # или какая-то заглушка, если 0 или 1 сэмпл
+    if b_samples_plot.size > 1: b_kde = gaussian_kde(b_samples_plot)
+    else: b_kde = None
+
+    a_dist = a_kde(x) if a_kde else np.zeros_like(x)
+    b_dist = b_kde(x) if b_kde else np.zeros_like(x)
+
+    if np.max(a_dist, initial=0) > 0: a_dist = a_dist / np.max(a_dist) # Added initial=0 for safety with empty arrays
+    if np.max(b_dist, initial=0) > 0: b_dist = b_dist / np.max(b_dist) # Added initial=0 for safety with empty arrays
+    kde_calc_end_time = time.time()
+    print(f"[LOG] KDE calculations (from N={N_SAMPLES_PLOTS} samples): {kde_calc_end_time - kde_calc_start_time:.4f}s")
 
     diff_dist_start_time = time.time()
-    diff_samples = b_samples - a_samples
-    # Ensure diff_samples is not empty before attempting np.min/np.max
-    if diff_samples.size > 0:
-        diff_x_min = np.min(diff_samples)
-        diff_x_max = np.max(diff_samples)
-        if diff_x_min == diff_x_max: # Avoids issue with linspace if min=max
-            diff_x_min -= 0.01
-            diff_x_max += 0.01
-        diff_x = np.linspace(diff_x_min, diff_x_max, 100)
-        kde_diff = gaussian_kde(diff_samples)
-        diff_distribution = kde_diff(diff_x)
-        if np.max(diff_distribution) > 0: diff_distribution = diff_distribution / np.max(diff_distribution) # Normalize
-    else: # Handle empty diff_samples case
-        diff_x = []
-        diff_distribution = []
+    # Создаем diff_samples_plot из _plot сэмплов
+    diff_samples_plot = b_samples_plot - a_samples_plot
+    # Используем diff_samples_plot для KDE
+    if diff_samples_plot.size > 1 :
+        diff_x_min = np.min(diff_samples_plot)
+        diff_x_max = np.max(diff_samples_plot)
+        # Обеспечить, что diff_x_min < diff_x_max для linspace и KDE
+        if diff_x_min >= diff_x_max:
+             diff_x_min = diff_x_min - 0.01 # Пример простого отступа
+             diff_x_max = diff_x_max + 0.01
+        if diff_x_min == diff_x_max: # если все значения одинаковы после отступа (маловероятно, но для защиты)
+             diff_x = np.array([diff_x_min])
+             diff_distribution = np.array([1.0])
+        else:
+             diff_x = np.linspace(diff_x_min, diff_x_max, 100)
+             kde_diff = gaussian_kde(diff_samples_plot)
+             diff_distribution = kde_diff(diff_x)
+    else: # если мало данных для KDE
+        diff_x = np.array([-0.1, 0, 0.1]) # какая-то заглушка
+        diff_distribution = np.array([0.0,1.0,0.0]) # какая-то заглушка
+
+    if np.max(diff_distribution, initial=0) > 0: diff_distribution = diff_distribution / np.max(diff_distribution) # Added initial=0 for safety
     diff_dist_end_time = time.time()
-    print(f"[LOG] Difference distribution calculation: {diff_dist_end_time - diff_dist_start_time:.4f}s")
+    print(f"[LOG] Difference distribution calculation (from N={N_SAMPLES_PLOTS} samples): {diff_dist_end_time - diff_dist_start_time:.4f}s")
 
     hist_calc_start_time = time.time()
     bins = 300 # Kept for now, might be used by new sample generation
-    # Ensure a_samples and b_samples are not empty before trying to create histograms
-    if a_samples.size > 0:
-        a_hist, bin_edges = np.histogram(a_samples, bins=bins, range=(0, 1), density=True)
+    # Используем _plot сэмплы для гистограмм
+    if a_samples_plot.size > 0:
+        a_hist, bin_edges = np.histogram(a_samples_plot, bins=bins, range=(0, 1), density=True)
         x_hist = 0.5 * (bin_edges[:-1] + bin_edges[1:])
     else:
-        a_hist = [] # Provide empty lists if no samples
-        x_hist = [] # Ensure x_hist is also empty if a_samples is empty
-        bin_edges = np.linspace(0,1,bins+1) # default bin_edges for x_hist if a_samples is empty
+        a_hist = np.zeros(bins).tolist() # заглушка
+        bin_edges = np.linspace(0,1,bins+1) # default bin_edges
+        x_hist = 0.5 * (bin_edges[:-1] + bin_edges[1:]) # x_hist based on default bins
 
-    if b_samples.size > 0:
-        b_hist, _ = np.histogram(b_samples, bins=bins, range=(0, 1), density=True)
-        if not x_hist.size > 0: # if a_samples was empty, x_hist would be empty
-             _, bin_edges_b = np.histogram(b_samples, bins=bins, range=(0, 1), density=True) # need bin_edges for x_hist
-             x_hist = 0.5 * (bin_edges_b[:-1] + bin_edges_b[1:])
+    if b_samples_plot.size > 0:
+        b_hist, _ = np.histogram(b_samples_plot, bins=bins, range=(0, 1), density=True)
     else:
-        b_hist = []
+        b_hist = np.zeros(bins).tolist() # заглушка
+
+    # Ensure x_hist is tolist() if it's a numpy array from calculation
+    if isinstance(x_hist, np.ndarray):
+        x_hist = x_hist.tolist()
+    if isinstance(a_hist, np.ndarray):
+        a_hist = a_hist.tolist()
+    if isinstance(b_hist, np.ndarray):
+        b_hist = b_hist.tolist()
 
     hist_calc_end_time = time.time()
-    print(f"[LOG] Histogram calculations: {hist_calc_end_time - hist_calc_start_time:.4f}s")
+    print(f"[LOG] Histogram calculations (from N={N_SAMPLES_PLOTS} samples): {hist_calc_end_time - hist_calc_start_time:.4f}s")
 
     processing_end_time = time.time() # This specific log can remain if data processing part is still meaningful
     print(f"[LOG] Data processing: Total {processing_end_time - processing_start_time:.4f}s") # This specific log can remain
@@ -221,9 +246,9 @@ def calculate_abtest(data: ABTestInput):
         b_expected_loss=round(b_expected_loss, 6),
         diff_x=diff_x.tolist(),
         diff_distribution=diff_distribution.tolist(),
-        a_hist=a_hist.tolist(),
-        b_hist=b_hist.tolist(),
-        x_hist=x_hist.tolist()
+        a_hist=a_hist, # Already a list
+        b_hist=b_hist, # Already a list
+        x_hist=x_hist  # Already a list
     )
     return_prep_end_time = time.time()
     print(f"[LOG] Result preparation: {return_prep_end_time - return_prep_start_time:.4f}s")