--- a/server_log_tiered_sampling.txt
+++ b/server_log_tiered_sampling.txt
@@ -1,0 +1,25 @@
+INFO:     Started server process [15452]
+INFO:     Waiting for application startup.
+INFO:     Application startup complete.
+INFO:     Uvicorn running on http://127.0.0.1:8000 (Press CTRL+C to quit)
+[LOG] calculate_abtest: Start
+[LOG] Frequentist calculations: 0.0005s
+[LOG] Bayesian part (Analytic Beta): Start
+[LOG] Posterior A: Beta(440.0, 22776.0)
+[LOG] Posterior B: Beta(495.0, 22602.0)
+[LOG] Sampling for metrics (N=1000000): 0.3686s
+[LOG] Slicing samples for plots (N=50000): 0.0000s
+[LOG] Bayesian part (Analytic Beta): Total 0.3687s
+[LOG] Data processing: Start
+[LOG] Bayesian metrics (from N=1000000 samples): 0.1074s
+[LOG] KDE calculations (from N=50000 samples): 0.5272s
+[LOG] Difference distribution calculation (from N=50000 samples): 0.0930s
+[LOG] Histogram calculations (from N=50000 samples): 0.0027s
+[LOG] Data processing: Total 0.7305s
+[LOG] Result preparation: 0.0003s
+[LOG] calculate_abtest: Total execution 1.1001s
+INFO:     127.0.0.1:43682 - "POST /calculate HTTP/1.1" 200 OK
+INFO:     Shutting down
+INFO:     Waiting for application shutdown.
+INFO:     Application shutdown complete.
+INFO:     Finished server process [15452]