--- a/server_log_user_data_2M.txt
+++ b/server_log_user_data_2M.txt
@@ -1,0 +1,24 @@
+INFO:     Started server process [14834]
+INFO:     Waiting for application startup.
+INFO:     Application startup complete.
+INFO:     Uvicorn running on http://127.0.0.1:8000 (Press CTRL+C to quit)
+[LOG] calculate_abtest: Start
+[LOG] Frequentist calculations: 0.0005s
+[LOG] Bayesian part (Analytic Beta): Start
+[LOG] Posterior A: Beta(440.0, 22776.0)
+[LOG] Posterior B: Beta(495.0, 22602.0)
+[LOG] Sampling from posterior Beta distributions (N=2000000): 0.9530s
+[LOG] Bayesian part (Analytic Beta): Total 0.9530s
+[LOG] Data processing: Start
+[LOG] KDE calculations (a_dist, b_dist): 23.5891s
+[LOG] Bayesian metrics (prob_b_better, means, expected_loss): 0.0882s
+[LOG] Difference distribution calculation: 4.3169s
+[LOG] Histogram calculations: 0.0736s
+[LOG] Data processing: Total 28.0681s
+[LOG] Result preparation: 0.0002s
+[LOG] calculate_abtest: Total execution 29.0218s
+INFO:     127.0.0.1:35752 - "POST /calculate HTTP/1.1" 200 OK
+INFO:     Shutting down
+INFO:     Waiting for application shutdown.
+INFO:     Application shutdown complete.
+INFO:     Finished server process [14834]